AL: Fixed bug for get character by id endpoint. Changed itn to long and re-structured pathvarable.

+++ src/main/java/com/example/assignmentwebapi/repository/CharacterRepository.java
@@ -11,7 +11,7 @@ import java.util.Optional;
 @Repository
 public interface CharacterRepository extends JpaRepository<Character, Integer> {
 
-Optional<Character> findCharacterByCharacterId(int characterId);
+Optional<Character> findCharacterByCharacterId(long characterId);
